PR: Support multiple [z_thermal_adjust] config sections (#751)
Multiple `z_thermal_adjust` sections may now be defined in config to compensate for different sources of z thermal expansion. E.g. the fram and nozzle expand at dfferent rates and time during the print. This now allows both to be modeled for more accurate z-adjustment behaviour.

There is an explanation and test data in this thread: https://klipper.discourse.group/t/z-thermal-adjust-for-nozzles/22457

Signed-off-by: Gareth Farrington <gareth@waves.ky>